refactor: reorganize default catalog module structure
Moving default catalog into separate crate
