+++ src/query/default_catalog/src/databases/database_context.rs
@@ -12,4 +12,23 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-pub use databend_query_default_catalog::databases::*;
+use std::sync::Arc;
+
+use databend_common_meta_app::tenant::Tenant;
+use databend_common_meta_store::MetaStore;
+use databend_common_storages_factory::StorageFactory;
+
+/// Database Context.
+#[derive(Clone)]
+pub struct DatabaseContext {
+    pub meta: MetaStore,
+    pub storage_factory: Arc<StorageFactory>,
+    pub tenant: Tenant,
+    pub disable_table_info_refresh: bool,
+}
+
+impl DatabaseContext {
+    pub fn tenant(&self) -> &Tenant {
+        &self.tenant
+    }
+}
