+++ src/query/service/tests/it/catalogs/immutable_catalogs.rs
@@ -20,8 +20,8 @@ use databend_common_meta_app::schema::CreateOption;
 use databend_common_meta_app::schema::DropDatabaseReq;
 use databend_common_meta_app::schema::RenameDatabaseReq;
 use databend_common_meta_app::tenant::Tenant;
-use databend_query::catalogs::default::ImmutableCatalog;
 use databend_query::catalogs::Catalog;
+use databend_query::catalogs::ImmutableCatalog;
 
 use crate::tests::create_catalog;
